fix: add JET static analysis tests, fix found issues and improve type stability
- Removed unused Header constructor for IO streams
- Added explicit imports for CommonDataModel interface methods

+++ src/loading/variable.jl
@@ -71,8 +71,8 @@ function DiskArrays.readblock!(var::CDFVariable{T, N}, dest::AbstractArray{T}, r
 
     first_rec = first(record_range)
     last_rec = last(record_range)
-    start_idx = findfirst(entry -> entry.first <= first_rec <= entry.last, entries)
-    end_idx = findfirst(entry -> entry.first <= last_rec <= entry.last, entries)
+    start_idx = findfirst(entry -> entry.first <= first_rec <= entry.last, entries)::Int
+    end_idx = findfirst(entry -> entry.first <= last_rec <= entry.last, entries)::Int
     record_size = prod(dims_without_record)
 
     is_row_major = majority(var) == Row
